requirements.txt
streamlit>=1.24
tensorflow
Pillow
numpy
gdown
